Refactoring the rules somewhat
This probably breaks the scripts, they are not debugged.
An additional column appears that allows for alternates (for a given item/subitem). This should be used as a selector in the pattern, eg. restud.csv, but it could also be used in a selector app to enforce that one of a set of alternates is chosen.